requirements.txt
# If your project has any requirements not included in anaconda, include them here
pandas_datareader